chore(desktop): satisfy line-count ratchet and flow coverage for the opener/ask-bar work
Moves the shelved FloatingPageAskBar into its own MainWindow file (keeps
DesktopHomeView under its ratchet threshold), raises the automation-bridge
baseline for the present_onboarding_opener QA action, ratchets the MainWindow
shard down for the DashboardPage shrink, and maps the new file to the
navigation e2e flow that owns page-level UI.
