+++ desktop/macos/e2e/flows/navigation.yaml
@@ -8,6 +8,7 @@ covers:
   - desktop/macos/Desktop/Sources/MainWindow/CrispManager.swift
   - desktop/macos/Desktop/Sources/MainWindow/SidebarView.swift
   - desktop/macos/Desktop/Sources/MainWindow/DesktopHomeView.swift
+  - desktop/macos/Desktop/Sources/MainWindow/FloatingPageAskBar.swift
   - desktop/macos/Desktop/Sources/MainWindow/Pages/HomeStatusStore.swift
   - desktop/macos/Desktop/Sources/Theme/OmiButtonStyle.swift
   - desktop/macos/Desktop/Sources/Theme/OmiChrome.swift
